Isolate scheduled task types
