+++ src/models/scheduled-task/is-scheduled-query.ts
@@ -1,0 +1,25 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+import { isBoolean, isNumber, isString } from 'lodash';
+import { isScheduledTaskBase } from './is-scheduled-task-base';
+import { ScheduledQuery } from './squeduled-query';
+
+export const isScheduledQuery = (value: any): value is ScheduledQuery => {
+	try {
+		const sq = <ScheduledQuery>value;
+		return (
+			isScheduledTaskBase(sq) &&
+			sq.type === 'query' &&
+			isString(sq.query) &&
+			(isNumber(sq.searchSince.secondsAgo) || isBoolean(sq.searchSince.lastRun))
+		);
+	} catch {
+		return false;
+	}
+};
